+++ Changelog.md
@@ -9,5 +9,6 @@
     - 16 new wool slabs
     - 16 new wool walls
     - 16 new wool buttons
+    - 16 new wool pressure plates
 - right clicking with shears will drop them instantly
 - right clicking with a dye will change the color of the block